WEB-3770 migrate buildGenerateAGPImages tests
